requirements/required.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==1.1.0
django-redis==5.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
drf-yasg==1.21.6
drf-access-policy==1.5.0
idna==3.4
fast-captcha==0.2.1
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
pkgutil_resolve_name==1.3.10
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyrsistent==0.19.3
pyparsing==3.1.1
pytz==2023.3
PyYAML==6.0
Pillow==9.5.0
redis==4.5.5
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0
six==1.16.0
